feat: speed up release pieces ci/cd
