+++ tools/scripts/pieces/update-pieces-metadata.ts
@@ -2,12 +2,11 @@ import assert from 'node:assert';
 import { PieceMetadata } from '../../../packages/pieces/community/framework/src';
 import { StatusCodes } from 'http-status-codes';
 import { HttpHeader } from '../../../packages/pieces/community/common/src';
-import { findAllPieces } from '../utils/piece-script-utils';
+import { AP_CLOUD_API_BASE, findAllPieces, findNewPieces, pieceMetadataExists } from '../utils/piece-script-utils';
 
 assert(process.env['AP_CLOUD_API_KEY'], 'API Key is not defined');
 
 const { AP_CLOUD_API_KEY } = process.env;
-const AP_CLOUD_API_BASE = 'https://cloud.activepieces.com/api/v1';
 
 const insertPieceMetadata = async (
   pieceMetadata: PieceMetadata
@@ -30,28 +29,7 @@ const insertPieceMetadata = async (
   }
 };
 
-const pieceMetadataExists = async (
-  pieceName: string,
-  pieceVersion: string
-): Promise<boolean> => {
-  const cloudResponse = await fetch(
-    `${AP_CLOUD_API_BASE}/pieces/${pieceName}?version=${pieceVersion}`
-  );
 
-  const pieceExist: Record<number, boolean> = {
-    [StatusCodes.OK]: true,
-    [StatusCodes.NOT_FOUND]: false
-  };
-
-  if (
-    pieceExist[cloudResponse.status] === null ||
-    pieceExist[cloudResponse.status] === undefined
-  ) {
-    throw new Error(await cloudResponse.text());
-  }
-
-  return pieceExist[cloudResponse.status];
-};
 
 const insertMetadataIfNotExist = async (pieceMetadata: PieceMetadata) => {
   console.info(
@@ -72,16 +50,14 @@ const insertMetadataIfNotExist = async (pieceMetadata: PieceMetadata) => {
 };
 
 const insertMetadata = async (piecesMetadata: PieceMetadata[]) => {
-  for (const pieceMetadata of piecesMetadata) {
-    await insertMetadataIfNotExist(pieceMetadata);
-  }
+  await Promise.all(piecesMetadata.map(insertMetadataIfNotExist))
 };
 
 
 const main = async () => {
   console.log('update pieces metadata: started')
 
-  const piecesMetadata = await findAllPieces()
+  const piecesMetadata = await findNewPieces()
   await insertMetadata(piecesMetadata)
 
   console.log('update pieces metadata: completed')
